common: add helpers for option building
